refactor: MyPageLayout에서 사용자 정보 로딩 로직을 MyPageClient로 이동하여 코드 구조 개선 (#112)
- MyPageLayout 컴포넌트에서 사용자 정보 로딩 및 상태 관리를 MyPageClient 컴포넌트로 분리하여 코드의 가독성과 유지보수성을 향상
- 로딩 상태 및 프로필 정보 처리를 MyPageClient에서 관리하도록 변경하여 컴포넌트 간 역할 분담 명확화
